requirements.txt
asgiref==3.3.4
autoflake==1.4
autopep8==1.5.6
Django==3.2
django-debug-toolbar==3.2.1
django-filter==2.4.0
django-taggit==1.3.0
djangorestframework==3.12.4
docformatter==1.4
isort==5.8.0
Markdown==3.3.4
Pillow==8.2.0
pycodestyle==2.7.0
pyflakes==2.3.1
pyformat==0.7
python-dotenv==0.17.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
unify==0.5
untokenize==0.1.1
whitenoise==5.2.0
